refactor[sgt::TextFormat, sgt::JsonFormat]: combine two classes into one file for easy access
