+++ source/formatter.h
@@ -1,9 +1,16 @@
 #pragma once
 #include "formatInterface.h"
+
 namespace sgt {
 	class JsonFormat : public sgt::FormatInterface {
 	public:
 		virtual ~JsonFormat() = default;
+		std::string format(const std::string& text, const LogLevel& level) override;
+	};
+
+	class TextFormat : public FormatInterface {
+	public:
+		virtual ~TextFormat() = default;
 		std::string format(const std::string& text, const LogLevel& level) override;
 	};
 }
